refactor: manually chunk components to prevent excessive build bundle
